ModernizeTypesCastingFixer: relocate util methods into an abstract classes (maintainers request)
