+++ Symfony/CS/AbstractFunctionReferenceFixer.php
@@ -9,15 +9,18 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Symfony\CS\Functions\FunctionReference;
+namespace Symfony\CS;
 
 use Symfony\CS\Tokenizer\Tokens;
 
-final class Finder
+/**
+ * @author Vladimir Reznichenko <kalessil@gmail.com>
+ */
+abstract class AbstractFunctionReferenceFixer extends AbstractFixer
 {
     /**
      * Looks up Tokens sequence for suitable candidates and delivers boundaries information,
-     * which can be supplied by Symfony\CS\Functions\*Util classes.
+     * which can be supplied by other methods in this abstract class.
      *
      * @param string $functionNameToSearch
      * @param Tokens $tokens
@@ -26,7 +29,7 @@ final class Finder
      *
      * @return int[]|null returns $functionName, $openParenthesis, $closeParenthesis packed into array
      */
-    public static function find($functionNameToSearch, Tokens $tokens, $start, $end)
+    protected function find($functionNameToSearch, Tokens $tokens, $start, $end)
     {
         // find raw sequence which we can analyse for context
         $candidateSequence = array(array(T_STRING, $functionNameToSearch), '(');
@@ -66,5 +69,55 @@ final class Finder
         }
 
         return array($functionName, $openParenthesis, $closeParenthesis);
+    }
+
+    /**
+     * Count amount of parameters in a function/method reference.
+     *
+     * @param int                                  $openParenthesis
+     * @param int                                  $closeParenthesis
+     * @param Tokens|\Symfony\CS\Tokenizer\Token[] $tokens
+     *
+     * @return int
+     */
+    protected function countArguments($openParenthesis, $closeParenthesis, Tokens $tokens)
+    {
+        $firstSensibleToken = $tokens->getNextMeaningfulToken($openParenthesis);
+        if (')' === $tokens[$firstSensibleToken]->getContent()) {
+            return 0;
+        }
+
+        $argumentsCount = 1;
+        for ($paramContentIndex = $openParenthesis + 1; $paramContentIndex < $closeParenthesis; ++$paramContentIndex) {
+            // skip nested (...) constructs
+            if ('(' === $tokens[$paramContentIndex]->getContent()) {
+                $paramContentIndex = $tokens->findBlockEnd(Tokens::BLOCK_TYPE_PARENTHESIS_BRACE, $paramContentIndex);
+                continue;
+            }
+
+            // if comma matched, increase arguments counter
+            if (',' === $tokens[$paramContentIndex]->getContent()) {
+                ++$argumentsCount;
+            }
+        }
+
+        return $argumentsCount;
+    }
+
+    /**
+     * Checks if function/method defined in the scope.
+     *
+     * @param string $functionName
+     * @param Tokens $tokens
+     *
+     * @return bool
+     */
+    protected function isDefinedInScope($functionName, Tokens $tokens)
+    {
+        $definitionSequence = array(array(T_FUNCTION, 'function'), array(T_STRING, $functionName));
+
+        $matchedDefinition = $tokens->findSequence($definitionSequence, 0, $tokens->count() - 1, false);
+
+        return null !== $matchedDefinition;
     }
 }
